pyproject.toml
[project]
name = "etl-2"
version = "0.1.0"
description = ""
requires-python = ">=3.10,<4.0"

dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.123.5",
    "pyhumps>=3.8.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn>=0.38.0",
]

[tool.uv]
dev-dependencies = [
    "pipdeptree==2.26.1",
]
